+++ src/components/headers/BarraDePesquisa.jsx
@@ -37,13 +37,13 @@ function BarraDePesquisa({ placeholder = 'Pesquisar...' }) {
 
   return (
     <div className="flex items-center bg-[#f5dfb6] border border-[#db520a] rounded-full sm:px-0 md:px-4 lg:px-4 py-2 w-[45%] sm:w-[45%] md:w-[80%] lg:w-full max-w-4xl">
-      <IoSearch alt="Pesquisar" color="#DB520A" className="w-[40px] h-[40px] hidden sm:block ml-4" />
+      <IoSearch alt="Pesquisar" color="#DB520A" className="w-[40px] h-[40px] hidden sm:block ml-4 flex-shrink-0" />
       <input
         type="search"
         ref={inputRef}
         value={searchValue}
         placeholder={placeholder}
-        className="w-full bg-transparent border-none outline-none font-poppins sm:text-[20px] md:text-[24px] lg:text-[32px] text-[#1e1e1e] px-4"
+        className="w-full bg-transparent border-none outline-none font-poppins sm:text-[20px] md:text-[24px] lg:text-[24px] text-[#1e1e1e] px-4"
         onChange={handleInputChange}
       />
       {searchValue && (