+++ src/components/headers/Header2.jsx
@@ -8,12 +8,16 @@ import { FiFilter } from 'react-icons/fi';
 import BarraDePesquisa from './BarraDePesquisa';
 import { useAuth } from '../../features/auth/useAuth';
 import ModalAvaliacao from '../ModalAvaliacao';
+import ModalFiltroEstabelecimento from '../ModalFiltroEstabelecimento';
+import ModalFiltroAvaliacao from '../ModalFiltroAvaliacoes';
 
 //Padrão de Header com logo com título, barra de pesquisa expansível, botão de filtrar e botão de sair
-function Header({ placeholder }) {
+function Header({ placeholder, onAplicarFiltros, filtros, ordenador }) {
   const [mostrarPesquisa, setMostrarPesquisa] = useState(false);
   const [modalVisivel, setModalVisivel] = useState(false);
+  const [modalFiltroVisivel, setModalFiltroVisivel] = useState(false);
   const { logout } = useAuth();
+  const { user } = useAuth();
   const location = useLocation();
 
   const alternarPesquisa = () => {
@@ -21,6 +25,8 @@ function Header({ placeholder }) {
   };
 
   const mostrarBotaoNovaAvaliacao = location.pathname === '/feed';
+  const mostrarBotoesPesquisaEFiltro =
+    !location.pathname.startsWith(`/perfil/${user.username}`) && !location.pathname.startsWith('/estabelecimento/');
 
   return (
     <header
@@ -59,16 +65,39 @@ function Header({ placeholder }) {
       {mostrarPesquisa && <BarraDePesquisa placeholder={placeholder} />}
 
       <div className="flex gap-1 sm:gap-1 md:gap-4 lg:gap-4">
-        <button className="flex-shrink-0" onClick={alternarPesquisa}>
-          {mostrarPesquisa ? (
-            <IoClose className="w-[30px] sm:w-[30px] md:w-[40px] lg:w-[45px] h-auto" alt="Fechar pesquisa" />
-          ) : (
-            <IoSearch className="w-[30px] sm:w-[30px] md:w-[40px] lg:w-[45px] h-auto" alt="Pesquisar" />
-          )}
-        </button>
-        <button className="flex-shrink-0">
-          <FiFilter className="w-[30px] sm:w-[30px] md:w-[40px] lg:w-[45px] h-auto" alt="Filtrar" />
-        </button>
+        {mostrarBotoesPesquisaEFiltro && (
+          <>
+            <button className="flex-shrink-0" onClick={alternarPesquisa}>
+              {mostrarPesquisa ? (
+                <IoClose className="w-[30px] sm:w-[30px] md:w-[40px] lg:w-[45px] h-auto" alt="Fechar pesquisa" />
+              ) : (
+                <IoSearch className="w-[30px] sm:w-[30px] md:w-[40px] lg:w-[45px] h-auto" alt="Pesquisar" />
+              )}
+            </button>
+            <button className="flex-shrink-0" onClick={() => setModalFiltroVisivel(true)}>
+              <FiFilter className="w-[30px] sm:w-[30px] md:w-[40px] lg:w-[45px] h-auto" alt="Filtrar" />
+            </button>
+          </>
+        )}
+
+        {modalFiltroVisivel && location.pathname == '/estabelecimento' && (
+          <ModalFiltroEstabelecimento
+            open={modalFiltroVisivel}
+            onClose={() => setModalFiltroVisivel(false)}
+            onAplicar={onAplicarFiltros}
+            filtros={filtros}
+            ordenador={ordenador}
+          />
+        )}
+        {modalFiltroVisivel && location.pathname == '/feed' && (
+          <ModalFiltroAvaliacao
+            open={modalFiltroVisivel}
+            onClose={() => setModalFiltroVisivel(false)}
+            onAplicar={onAplicarFiltros}
+            filtros={filtros}
+            ordenador={ordenador}
+          />
+        )}
 
         {mostrarBotaoNovaAvaliacao && (
           <button