+++ src/router/routes.jsx
@@ -3,14 +3,16 @@ import Login from '../pages/Login';
 import NaoEncontrado from '../pages/NaoEncontrado';
 import PrivateRoute from './PrivateRoute';
 import PublicRoute from './PublicRoute';
-import Layout4 from '../components/layouts/Layout4';
+import Feed from '../pages/Feed';
 import FeedEstabelecimentos from '../pages/FeedEstabelecimentos';
 import RecuperarSenha from '../pages/RecuperarSenha';
 import NovaSenha from '../pages/NovaSenha';
 import CriarConta from '../pages/CriarConta';
+import EditarPerfil from '../pages/EditarPerfil';
 import Perfil from '../pages/Perfil';
 import EstabelecimentosFiltrados from '../pages/EstabelecimentosFiltrados';
 import UsuariosFiltrados from '../pages/UsuarioFiltrados';
+import Estabelecimento from '../pages/Estabelecimento';
 
 export default function RoutesApp() {
   return (
@@ -30,7 +32,7 @@ export default function RoutesApp() {
         path="/feed"
         element={
           <PrivateRoute>
-            <Layout4 />
+            <Feed />
           </PrivateRoute>
         }
       />
@@ -40,6 +42,15 @@ export default function RoutesApp() {
         element={
           <PrivateRoute>
             <FeedEstabelecimentos />
+          </PrivateRoute>
+        }
+      />
+
+      <Route
+        path="/estabelecimento/:id"
+        element={
+          <PrivateRoute>
+            <Estabelecimento />
           </PrivateRoute>
         }
       />
@@ -69,9 +80,12 @@ export default function RoutesApp() {
           <PrivateRoute>
             {' '}
             <UsuariosFiltrados />
+        path="/perfil/editar"
+        element={
+          <PrivateRoute>
+            <EditarPerfil />
           </PrivateRoute>
-        }
-      />
+        }</PrivateRoute>} />
 
       <Route path="/recuperar-senha" element={<RecuperarSenha />} />
       <Route path="/alterar-senha" element={<NovaSenha />} />