+++ CHANGELOG.md
@@ -63,3 +63,39 @@
 ### Adicionando
 
 - Conexão de botão de editar avaliação com modal de edição
+
+## US7.4 - PR 16
+
+### Adicionando
+
+- Tela de exibir estabelecimento
+
+## US6.4 - PR 11
+
+### Adicionando
+
+- Modal de detalhar avaliações
+
+## US5.4 - PR 17
+
+### Adicionando
+
+- Tela de edição/exclusão de perfil
+
+## US6.2 - PR 18
+
+### Adicionando
+
+- Tela do feed
+
+## US8.2 - PR 20
+
+### Adicionando
+
+- Modal filtro de avaliações
+
+## US8.4 - PR 19
+
+### Adicionando
+
+- Modal filtro de estabelecimentos